* [refac bug #69666] Adjust for sorna.

+++ module/task/model.php
@@ -194,32 +194,29 @@ class taskModel extends model
                 $this->loadModel('common')->syncPPEStatus($oldTask->id);
             }
 
-            if($this->config->edition != 'open')
+            if($this->config->edition != 'open' && $task->story != $oldTask->story)
             {
-                if($task->story != $oldTask->story)
+                if($oldTask->story > 0)
                 {
-                    if($oldTask->story > 0)
-                    {
-                        $this->dao->delete()->from(TABLE_RELATION)
-                            ->where('relation')->eq('generated')
-                            ->andWhere('AID')->eq($oldTask->story)
-                            ->andWhere('AType')->eq('story')
-                            ->andWhere('BID')->eq($oldTask->id)
-                            ->andWhere('BType')->eq('task')
-                            ->exec();
-                    }
+                    $this->dao->delete()->from(TABLE_RELATION)
+                        ->where('relation')->eq('generated')
+                        ->andWhere('AID')->eq($oldTask->story)
+                        ->andWhere('AType')->eq('story')
+                        ->andWhere('BID')->eq($oldTask->id)
+                        ->andWhere('BType')->eq('task')
+                        ->exec();
+                }
 
-                    if($task->story > 0)
-                    {
-                        $relation = new stdClass();
-                        $relation->relation = 'generated';
-                        $relation->AID      = $task->story;
-                        $relation->AType    = 'story';
-                        $relation->BID      = $oldTask->id;
-                        $relation->BType    = 'task';
-                        $relation->product  = 0;
-                        $this->dao->replace(TABLE_RELATION)->data($relation)->exec();
-                    }
+                if($task->story > 0)
+                {
+                    $relation = new stdClass();
+                    $relation->relation = 'generated';
+                    $relation->AID      = $task->story;
+                    $relation->AType    = 'story';
+                    $relation->BID      = $oldTask->id;
+                    $relation->BType    = 'task';
+                    $relation->product  = 0;
+                    $this->dao->replace(TABLE_RELATION)->data($relation)->exec();
                 }
             }
         }
